Fixed & Refactored Login/Signup Forms
Removed the need for inputRefs. Cleaned up the code and changed some wording.
